pyproject.toml
[project]
name = "augint-github"
version = "1.11.0"
description = "GitHub tools for Augmenting Integrations"
authors = [{name = "Samuel Vange", email = "7166607+svange@users.noreply.github.com"}]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12,<4.0"
dependencies = [
    "click>=8.1.8,<9",
    "python-dotenv>=1.0.1,<2",
    "rich>=13.9.4,<15.0",
    "loguru>=0.7.3,<0.8",
    "pygithub>=2.5.0,<3",
]

[project.urls]
Homepage = "https://github.com/svange/augint-github"

[project.scripts]
ai-gh = "gh_secrets_and_vars_async.cli:main"
ai-gh-push = "gh_secrets_and_vars_async:cli"

[project.optional-dependencies]
tui = ["textual>=1.0.0,<2"]
dev = [
    "textual>=1.0.0,<2",
    "pytest>=8.3.4",
    "pytest-html>=4.1.1",
    "pytest-cov>=6.0.0",
    "augint-shell>=0.16.7",
    "python-semantic-release>=10.3.1",
    "pre-commit>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "pip-audit>=2.7.0",
    "pip-licenses>=5.0.0",
    "pdoc>=15.0.0",
]

[build-system]
requires = ["uv_build>=0.11,<0.12"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "gh_secrets_and_vars_async"

[tool.semantic_release]
assets = ["uv.lock"]
commit_message = "chore(release): {version}\n\nAutomatically generated by python-semantic-release [skip ci]"
commit_parser = "conventional"
major_on_zero = false
tag_format = "v{version}"
version_toml = ["pyproject.toml:project.version"]
build_command = "uv lock && uv build"

[tool.semantic_release.branches.main]
match = "main"
prerelease = false

[tool.semantic_release.commit_parser_options]
patch_tags = ["fix", "perf"]

[tool.semantic_release.changelog]
mode = "update"
exclude_commit_patterns = [
    '''chore(?:\([^)]*?\))?: .+''',
    '''ci(?:\([^)]*?\))?: .+''',
    '''refactor(?:\([^)]*?\))?: .+''',
    '''style(?:\([^)]*?\))?: .+''',
    '''test(?:\([^)]*?\))?: .+''',
    '''build\((?!deps\): .+)''',
]

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = true
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = false

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    'pragma: no cover',
    'def __repr__',
    'if self\.debug',
    'raise AssertionError',
    'raise NotImplementedError',
    'if 0:',
    'if __name__ == .__main__.:',
    'if TYPE_CHECKING:',
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --strict-markers"
markers = [
    "redundant: mark test as redundant",
    "end_to_end: mark test as an end-to-end test",
    "no_infra: mark test as not requiring infrastructure",
    "skip_ci: mark as not suitable for ci",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "C4", "UP", "DTZ"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["gh_secrets_and_vars_async"]

[tool.mypy]
python_version = "3.12"
files = "src"
disallow_untyped_defs = true
disallow_untyped_decorators = true
disallow_subclassing_any = true
warn_return_any = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "gh_secrets_and_vars_async.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "github.*"
ignore_missing_imports = true
